requirements-prod.txt
caldav==3.0.0a1
celery==5.6.2
email-validator==2.3.0
eventlet==0.40.4
Flask==3.1.3
Flask-Admin==2.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
importlib_metadata==8.7.1
jinja_partials==0.3.1
Paste==3.10.1
PySocks==1.7.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
redis==7.2.1
sqlalchemy_schemadisplay==2.0
tzlocal==5.3.1
waitress==3.0.2

